+++ page_iterator.go
@@ -79,7 +79,7 @@ func NewPageIterator(res interface{}, reqAdapter abstractions.RequestAdapter, co
 //          return true
 //      }
 //      err := pageIterator.Iterate(callbackFunc)
-func (pI *PageIterator) Iterate(callback func(pageItem interface{}) bool) error {
+func (pI *PageIterator) Iterate(context context.Context, callback func(pageItem interface{}) bool) error {
 	for {
 		keepIterating := pI.enumerate(callback)
 
@@ -88,7 +88,7 @@ func (pI *PageIterator) Iterate(callback func(pageItem interface{}) bool) error 
 			return nil
 		}
 
-		nextPage, err := pI.next()
+		nextPage, err := pI.next(context)
 		if err != nil {
 			return err
 		}
@@ -110,14 +110,14 @@ func (pI *PageIterator) SetReqOptions(reqOptions []abstractions.RequestOption) {
 	pI.reqOptions = reqOptions
 }
 
-func (pI *PageIterator) next() (PageResult, error) {
+func (pI *PageIterator) next(context context.Context) (PageResult, error) {
 	var page PageResult
 
 	if pI.currentPage.getNextLink() == nil {
 		return page, nil
 	}
 
-	resp, err := pI.fetchNextPage()
+	resp, err := pI.fetchNextPage(context)
 	if err != nil {
 		return page, err
 	}
@@ -130,7 +130,7 @@ func (pI *PageIterator) next() (PageResult, error) {
 	return page, nil
 }
 
-func (pI *PageIterator) fetchNextPage() (serialization.Parsable, error) {
+func (pI *PageIterator) fetchNextPage(context context.Context) (serialization.Parsable, error) {
 	var graphResponse serialization.Parsable
 	var err error
 
@@ -149,7 +149,7 @@ func (pI *PageIterator) fetchNextPage() (serialization.Parsable, error) {
 	requestInfo.Headers = pI.headers
 	requestInfo.AddRequestOptions(pI.reqOptions)
 
-	graphResponse, err = pI.reqAdapter.SendAsync(context.Background(), requestInfo, pI.constructorFunc, nil, nil)
+	graphResponse, err = pI.reqAdapter.SendAsync(context, requestInfo, pI.constructorFunc, nil, nil)
 	if err != nil {
 		return graphResponse, errors.New("fetching next page failed")
 	}